Allow optional params to override callback
Previously, to override the OAuth callback URI per request you had to
associate the preferred URI into the consumer, which was a bit clumsy.
With this change you can pass "oauth_callback" to both
`request-token-request` and `authorization-url` via the optional
`params` parameter.

    (request-token-request
     consumer
     {"oauth_callback" "http://example.com/callback?state=csrf-token"})

    (authorization-url
     consumer
     {"oauth_callback" "http://example.com/callback?state=csrf-token"})

This is important when you have any state you want to preserve, because
unlike OAuth 2.0 there is no `state` parameter you pass to the provider.

Make sure to pass the preferred callback URI to both functions. If you
use a "request token" with the wrong callback URI you're authorisation
URL will not redirect the user back to your app as expected with most
providers (certainly Twitter).